README.md
In this Kernel, I'm going to train a linear regression model to predict the miles per gallon (mpg) of a given car using the scikit learn module in python. The scikit learn module in python is basically used to perform machine learning tasks. It contains a number of machine learning models and algorithms that are very good and helpful to use. Among these models are the regression models that are used to perform both simple and multiple linear regression.
The data we will be using for our regression analysis comprises of some technical specs of some cars. The dataset is downloaded from UCI Machine Learning Repository.The UCI Machine Learning Repository is a machine learning repository that contains free data you can use for your machine learning projects.

About the dataset:
The dataset is downloaded from UCI Machine Learning Repository. The link to the data is: 
https://archive.ics.uci.edu/ml/datasets/auto+mpg

Content:
Title: Auto-Mpg Data
Source: This dataset was taken from the StatLib library which is maintained at Carnegie Mellon University. The dataset was used in the 1983 American Statistical Association Exposition
Relevant information: The data concerns fuel consumption in miles per gallon, which needs to be predicted. The number of instances are 398 and the number of attributes are 9 including mpg attribute. The attributes information: (i) mpg - Miles Per Galon (ii) cylinders - the power unit of the car where gasoline is turned into power (iii) displacement - engine displacement of the car (iv) horsepower - rate of the engine performance (v) weight - the weight of a car (vi) acceleration - the acceleration of a car (vii) model - model of the car (viii) origin - the origin of the car (ix) car - the name of the car

Steps Performed:
(1) Imported and preprocessed the data.
(2) Applied heatmap to see the correlation present between different attributes.
(3) Splitting of data into test and train sets using scikit-learn libraries.
(4) Model is trained using Linear Regression, calculation of r2_score ; which helps us to see how accurately our model has been trained.
(5) Dumped the model using pickle for deployment.